server/requirements.txt
aiohttp==3.10.5
av==13.0.0
fastapi==0.115.0
#ffmpeg==1.4
ffmpeg_python==0.2.0
firebase_admin==6.5.0
ImageHash==4.3.1
# librosa==0.10.2.post1
librosa==0.9.2
numpy>=1.23.5,<2.0.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.4.0
pydantic==2.9.2
pydub==0.25.1
python-dotenv==1.0.1
scipy==1.13.1
tensorflow==2.17.0
torch==2.4.1
transformers==4.44.2
uvicorn==0.30.6
numba==0.56.4